further developped l bonus but might be more complicated than anticipated

+++ pf_extractor.c
@@ -14,34 +14,34 @@
 
 long long           idextract(t_parg *parg, va_list args)
 {
-    long long   nb;
+	long long   nb;
 
-    if (parg->mh == 1)
+	if (parg->mh == 1)
 		nb = (short)va_arg(args, int);
 	else if (parg->mh >= 2)
 		nb = (signed char)va_arg(args, int);
-    else if (parg->ml == 1)
-        nb = (long)va_arg(args, long);
-    else if (parg->ml >= 2)
-        nb = (long long)va_arg(args, long long);
+	else if (parg->ml == 1)
+		nb = (long)va_arg(args, int);
+	else if (parg->ml >= 2)
+		nb = (long long)va_arg(args, int);
 	else
 		nb = (int)va_arg(args, int);
-    return (nb);
+	return (nb);
 }
 
 unsigned long long  uxtract(t_parg *parg, va_list args)
 {
-    unsigned long long  nb;
+	unsigned long long  nb;
 
-    if (parg->mh == 1)
+	if (parg->mh == 1)
 		nb = (unsigned short)va_arg(args, int);
 	else if (parg->mh >= 2)
 		nb = (unsigned char)va_arg(args, int);
-    else if (parg->ml == 1)
-        nb = (unsigned long)va_arg(args, long);
-    else if (parg->ml >= 2)
-        nb = (unsigned long long)va_arg(args, long long);
+	else if (parg->ml == 1)
+		nb = (unsigned long)va_arg(args, int);
+	else if (parg->ml >= 2)
+		nb = (unsigned long long)va_arg(args, int);
 	else
 		nb = (unsigned int)va_arg(args, int);
-    return (nb);
+	return (nb);
 }
